move size_field func into common
